bugfix: handle decimal hours properly ticket-213

+++ backend/src/main/java/ch/time2log/backend/api/rest/dto/outbound/MemberActivityRecordDto.java
@@ -2,6 +2,7 @@ package ch.time2log.backend.api.rest.dto.outbound;
 
 import ch.time2log.backend.domain.models.MemberActivityRecord;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.List;
 import java.util.UUID;
@@ -11,7 +12,7 @@ public record MemberActivityRecordDto(
         String entryDate,
         UUID curriculumActivityId,
         String activityLabel,
-        int hours,
+        BigDecimal hours,
         String notes,
         Integer rating,
         UUID teamId,
